fix tests
* Fix circular import error
* Update conversational-agent prompt
* Add conversational-agent-without-tools to legacy list
